requirements.txt
biopython==1.77
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.5
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
